Refactoring: code organization
