+++ package.json
@@ -21,7 +21,7 @@
     "onCommand:extension.elixirRunTestSuite",
     "onCommand:extension.elixirRunTestAtCursor"
   ],
-  "main": "./extension.js",
+  "main": "./src/extension.js",
   "contributes": {
     "commands": [
       {
