+++ src/commands/index.js
@@ -1,0 +1,6 @@
+const jumpToTest = require("./jumpToTest");
+const runTestAtCursor = require("./runTestAtCursor");
+const runTestFile = require("./runTestFile");
+const runTestSuite = require("./runTestSuite");
+
+module.exports = [jumpToTest, runTestAtCursor, runTestFile, runTestSuite];
